% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretPredict.R
\name{aggregate_mean_or_first}
\alias{aggregate_mean_or_first}
\title{Aggregate mean or first}
\usage{
aggregate_mean_or_first(x)
}
\arguments{
\item{x}{a train object}
}
\value{
a data.table::data.table with predictions
}
\description{
For numeric data take the mean.  For character data take the first value.
}
\keyword{internal}
