% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{as.caretList.default}
\alias{as.caretList.default}
\title{Convert object to caretList object - For Future Use}
\usage{
\method{as.caretList}{default}(object)
}
\arguments{
\item{object}{R object}
}
\value{
NA
}
\description{
Converts object into a caretList  - For Future Use
}

