% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{as.caretList.list}
\alias{as.caretList.list}
\title{Convert list to caretList object}
\usage{
\method{as.caretList}{list}(list)
}
\arguments{
\item{list}{of caret models}
}
\value{
a \code{\link{caretList}} object
}

