% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{as.caretList}
\alias{as.caretList}
\title{Coerce object to caretList object}
\usage{
as.caretList(object)
}
\arguments{
\item{list}{of caret models}
}
\value{
a \code{\link{caretList}} object
}
\description{
Coerces list of caret models into a caretList
}

