% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class
#}
\name{caretEnsemble}
\alias{caretEnsemble}
\title{Class "caretEnsemble" of ensembled train objects from the caret package
#}
\usage{
caretEnsemble(all.models, optFUN = NULL, ...)
}
\arguments{
\item{all.models}{a list of caret models to ensemble.}

\item{optFUN}{the optimization function to use}

\item{...}{additional arguments to pass to the optimization function}
}
\value{
a \code{\link{caretEnsemble}} object
}
\description{
Ensembled model from input train objects.
#
#
}
\details{
#

#The object has the following items
#\itemize{
#\item{models - a list of the original models to be ensembled}
#\item{weights - a matrix with the weights for each model}
#\item{error - the final accuracy metric of the ensembled models}
#}
#
}
\section{Objects from the Class}{
 Objects are created by calls to
#\code{\link{caretEnsemble}}.
#
}
\references{
\url{http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.60.2859&rep=rep1&type=pdf}
}
\seealso{
\code{\link{caretEnsemble}}
#
}
\keyword{#}
\keyword{classes}

