% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{caretModelSpec}
\alias{caretModelSpec}
\title{Generate a specification for fitting a caret model}
\usage{
caretModelSpec(
  method = "rf",
  ...,
  allowed_models = unique(caret::modelLookup()$model)
)
}
\arguments{
\item{method}{the modeling method to pass to caret::train. A character vector for
built in models or a list for a custom model.  Will be passed to caret::train(method=x).}

\item{...}{Other arguments that will eventually be passed to caret::train}

\item{all_models}{a character vector representing the allowed built-in models.
By default uses all_models = unique(caret::modelLookup()$model)}
}
\value{
a list of lists
}
\description{
A caretModelSpec is a named list of arguments to be passed to caret::train.
This constructor also does some validation. Nothing is modified.
}
\examples{
caretModelSpec("rf", tuneLength = 5, preProcess = "ica")
}
