% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_bestpreds_indexes}
\alias{check_bestpreds_indexes}
\title{Check row indexes}
\usage{
check_bestpreds_indexes(modelLibrary)
}
\arguments{
\item{modelLibrary}{a list of predictins from caret models}
}
\description{
Check that the row indexes from a caretList are valid
}

