% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_bestpreds_obs}
\alias{check_bestpreds_obs}
\title{Check observeds}
\usage{
check_bestpreds_obs(modelLibrary)
}
\arguments{
\item{modelLibrary}{a list of predictins from caret models}
}
\description{
Check that a list of observed values from a caretList are valid
}

