% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_bestpreds_preds}
\alias{check_bestpreds_preds}
\title{Check predictions}
\usage{
check_bestpreds_preds(modelLibrary)
}
\arguments{
\item{modelLibrary}{a list of predictins from caret models}
}
\description{
Check that a list of predictions from a caretList are valid
}

