% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_bestpreds_resamples}
\alias{check_bestpreds_resamples}
\title{Check resamples}
\usage{
check_bestpreds_resamples(modelLibrary)
}
\arguments{
\item{modelLibrary}{a list of predictins from caret models}
}
\description{
Check that the resamples from a caretList are valid
}
