% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_binary_classification}
\alias{check_binary_classification}
\title{Check binary classification}
\usage{
check_binary_classification(list_of_models)
}
\arguments{
\item{list_of_models}{a list of caret models to check}
}
\description{
Check that the problem is a binary classification problem
}
