% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_caretList_classes}
\alias{check_caretList_classes}
\title{Checks caretList model classes}
\usage{
check_caretList_classes(list_of_models)
}
\arguments{
\item{list_of_models}{a list of caret models to check}
}
\description{
This function checks caretList classes
}

