% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_caretList_model_types}
\alias{check_caretList_model_types}
\title{Checks that caretList models are all of the same type.}
\usage{
check_caretList_model_types(list_of_models)
}
\arguments{
\item{list_of_models}{a list of caret models to check}
}
\description{
Checks that caretList models are all of the same type.
}

