% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretStack.R
\name{check_caretStack}
\alias{check_caretStack}
\title{Check caretStack object}
\usage{
check_caretStack(object)
}
\arguments{
\item{object}{a caretStack object}
}
\description{
Make sure a caretStack has both a caretList and a train object
}
\keyword{internal}
