% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_multiclass_excluded_level}
\alias{check_multiclass_excluded_level}
\title{Check multiclass excluded level}
\usage{
check_multiclass_excluded_level(excluded_level, num_classes)
}
\arguments{
\item{excluded_level}{the level to exclude}

\item{num_classes}{the number of classes}
}
\description{
Verifies that the multiclass excluded level is
within the range of the number of classes.
}
