% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretStack.R
\name{check_newdata_features}
\alias{check_newdata_features}
\title{Check original_features_included and newdata}
\usage{
check_newdata_features(newdata, original_features_included)
}
\arguments{
\item{newdata}{a data.frame of new data}

\item{original_features_included}{a character vector of original features used in training}
}
\description{
Make sure newdata contains all the original features used in training
}
\keyword{internal}
