% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretStack.R
\name{check_use_original_features}
\alias{check_use_original_features}
\title{Check the parameters use_original_features, new_X, and which.var are consistent}
\usage{
check_use_original_features(use_original_features, new_X, which.var)
}
\arguments{
\item{use_original_features}{a logical indicating whether to include the original features in the stack.}

\item{new_X}{a data.frame of new data to train the stack on.}

\item{which.var}{a vector indicating which original features to include in the stack.}
}
\description{
Check the parameters use_original_features, new_X, and which.var are consistent
when training a caretStack.
}
\keyword{internal}
