% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkCaretList}
\alias{checkCaretList}
\title{Run a series of checks on a caretList object}
\usage{
checkCaretList(list_of_models)
}
\arguments{
\item{list_of_models}{a list of caret models to check}
}
\description{
Basically, this function validates that a caretList object is in good shape and is ready to be ensembled by caretList or caretEnsemble.
}

