% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{checkCustomModel}
\alias{checkCustomModel}
\title{Validate a custom caret model info list}
\usage{
checkCustomModel(x)
}
\arguments{
\item{x}{a model info list (e.g. \code{getModelInfo("rf", regex=F)\[[1]]})}
}
\value{
validated model info list (i.e. x)
}
\description{
Currently, this only ensures that all model info lists
were also assigned a "method" attribute for consistency with usage
of non-custom models
}
