% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkModelClasses}
\alias{checkModelClasses}
\title{Checks caretList model classes}
\usage{
checkModelClasses(list_of_models)
}
\arguments{
\item{list_of_models}{a list of caret models to check}
}
\description{
This function checks that the models in a caretList are all of class train and have the same model type (classification or regression).  It also checks that classification models have saved probabilities.
}

