\name{checkModels_extractTypes}
\alias{checkModels_extractTypes}
\title{Check that a list of models are all train objects and are ready to be ensembled together}
\usage{
  checkModels_extractTypes(list_of_models, ...)
}
\arguments{
  \item{list_of_models}{a list of caret models to check}
}
\description{
  Check that a list of models are all train objects and are
  ready to be ensembled together
}

