\name{checkPred}
\alias{checkPred}
\title{Check that a list of predictions from caret models are all valid}
\usage{
  checkPred(list_of_models, ...)
}
\arguments{
  \item{list_of_models}{a list of caret models to check}
}
\description{
  Check that a list of predictions from caret models are
  all valid
}

