% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkPreds}
\alias{checkPreds}
\title{Check predictions}
\usage{
checkPreds(list_of_models)
}
\arguments{
\item{list_of_models}{a list of caret models to check}
}
\description{
Check that a list of predictions from caret models are all valid
}

