% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkRowIndexes}
\alias{checkRowIndexes}
\title{Check row indexes}
\usage{
checkRowIndexes(list_of_models)
}
\arguments{
\item{list_of_models}{a list of caret models to check}
}
\description{
Check that the row indexes from a caretList are valid
}

