% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{defaultMetric}
\alias{defaultMetric}
\title{Construct a default metric}
\usage{
defaultMetric(is_class, is_binary)
}
\arguments{
\item{is_class}{logical, is this a classification or regression problem.}

\item{is_binary}{logical, is this binary classification.}
}
\description{
Caret defaults to RMSE for classification and RMSE for regression.
For classification, I would rather use ROC.
}
