% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dotplot.caretStack}
\alias{dotplot.caretStack}
\title{Comparison dotplot for a caretStack object}
\usage{
\method{dotplot}{caretStack}(x, data = NULL, ...)
}
\arguments{
\item{x}{An object of class caretStack}

\item{data}{passed to dotplot}

\item{...}{passed to dotplot}
}
\description{
This is a function to make a dotplot from a caretStack.
}

