% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretPredict.R
\name{dropExcludedClass}
\alias{dropExcludedClass}
\title{Drop Excluded Class}
\usage{
dropExcludedClass(x, all_classes, excluded_class_id)
}
\arguments{
\item{x}{a data.table of predictions}

\item{all_classes}{a character vector of all classes}

\item{excluded_class_id}{an integer indicating the class to exclude}
}
\description{
Drop the excluded class from a prediction data.table
}
\keyword{internal}
