% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extractBestPreds}
\alias{extractBestPreds}
\title{Extract the best predictions from a train object}
\usage{
extractBestPreds(x)
}
\arguments{
\item{x}{a train object}
}
\description{
Extract predictions for the best tune from a model
}
