% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extractBestPredsAndObs}
\alias{extractBestPredsAndObs}
\title{Extract the best predictions (and observeds) from a list of train objects}
\usage{
extractBestPredsAndObs(list_of_models, excluded_class_id = 1L)
}
\arguments{
\item{list_of_models}{an object of class caretList or a list of caret models}

\item{excluded_class_id}{an integer indicating the class to exclude for classification models}
}
\description{
Extract predictions (and observeds) for the best tune from a list of caret models.
This function extracts the raw preds from regression models and the class probs from classification models.
Note that it extract preds and obs in one go, rather than separately. This is because caret can save the internal
preds/obs from all resamples rather than just the final.  So we subset the internal pred/obs to just the best tuning
(from caret) and return the pred and obs for that tune.
}
