% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/caretList.R
\name{extractCaretTarget.default}
\alias{extractCaretTarget.default}
\title{Extracts the target variable from a set of arguments headed to the caret::train.default function.}
\usage{
\method{extractCaretTarget}{default}(x, y, ...)
}
\arguments{
\item{x}{an object where samples are in rows and features are in columns. This could be a simple matrix, data frame or other type (e.g. sparse matrix). See Details below.}

\item{y}{a numeric or factor vector containing the outcome for each sample.}

\item{...}{ignored}
}
\description{
This function extracts the y variable from a set of arguments headed to a caret::train.default model.
}

