% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{extractCaretTarget.formula}
\alias{extractCaretTarget.formula}
\title{Extracts the target variable from a set of arguments headed to the caret::train.formula function.}
\usage{
\method{extractCaretTarget}{formula}(form, data, ...)
}
\arguments{
\item{form}{A formula of the form y ~ x1 + x2 + ...}

\item{data}{Data frame from which variables specified in formula are preferentially to be taken.}

\item{...}{ignored}
}
\description{
This function extracts the y variable from a set of arguments headed to a caret::train.formula model.
}
