% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{extractMetric.caretList}
\alias{extractMetric.caretList}
\title{Extract accuracy metrics from a \code{\link[caretEnsemble]{caretList}} object}
\usage{
\method{extractMetric}{caretList}(x, ...)
}
\arguments{
\item{x}{a caretList object}

\item{...}{passed to extractMetric.train}
}
\value{
A data.table with metrics from each model.
}
\description{
Extract the cross-validated accuracy metrics from each model in a caretList.
}
