% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretStack.R
\name{extractMetric.caretStack}
\alias{extractMetric.caretStack}
\title{Extract accuracy metrics from a \code{\link[caretEnsemble]{caretStack}} object}
\usage{
\method{extractMetric}{caretStack}(x, ...)
}
\arguments{
\item{x}{a caretStack object}

\item{...}{passed to extractMetric.train and extractMetric.caretList}
}
\value{
A data.table with metrics from the ensemble model and individual models.
}
\description{
Extract the cross-validated accuracy metrics from the ensemble model
and individual models in a caretStack.
}
