% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretPredict.R
\name{extractMetric}
\alias{extractMetric}
\title{Generic function to extract accuracy metrics from various model objects}
\usage{
extractMetric(x, ...)
}
\arguments{
\item{x}{An object from which to extract metrics.
The specific method will be dispatched based on the class of \code{x}.}

\item{...}{Additional arguments passed to the specific methods.}
}
\value{
A \code{\link[data.table]{data.table}}
}
\description{
A generic function to extract cross-validated accuracy metrics from model objects.
}
\seealso{
\code{\link{extractMetric.train}},
\code{\link{extractMetric.caretList}},
\code{\link{extractMetric.caretStack}}
}
