% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretPredict.R
\name{extractMetric.train}
\alias{extractMetric.train}
\title{Extract accuracy metrics from a \code{\link[caret]{train}} model}
\usage{
\method{extractMetric}{train}(x, metric = NULL, ...)
}
\arguments{
\item{x}{a train object}

\item{metric}{a character string representing the metric to extract.}

\item{...}{ignored
If NULL, uses the metric that was used to train the model.}
}
\value{
A numeric representing the metric desired metric.
}
\description{
Extract the cross-validated accuracy metrics and their SDs from caret.
}
