% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretEnsemble.R
\name{extractModelMetrics}
\alias{extractModelMetrics}
\title{Extract the model accuracy metrics of the individual models in an ensemble object.}
\usage{
extractModelMetrics(ensemble, metric = NULL)
}
\arguments{
\item{ensemble}{a caretEnsemble to make predictions from.}

\item{metric}{a character string representing the metric to extract.
If NULL, each model will return the metric it was trained on.
If not NULL, the specified metric must be present for EVERY trained model.}
}
\description{
Extract the model accuracy metrics of the individual models in an ensemble object.
}
