% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretPredict.R
\name{extractModelName}
\alias{extractModelName}
\title{Extract the method name associated with a single train object}
\usage{
extractModelName(x)
}
\arguments{
\item{x}{a single caret train object}
}
\value{
Name associated with model
}
\description{
Extracts the method name associated with a single train object. Note
that for standard models (i.e. those already prespecified by caret), the
"method" attribute on the train object is used directly while for custom
models the "method" attribute within the model$modelInfo attribute is
used instead.
}
