% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretPredict.R
\name{extractModelType}
\alias{extractModelType}
\title{Extract the model type from a \code{\link[caret]{train}} object}
\usage{
extractModelType(object, validate_for_stacking = TRUE)
}
\arguments{
\item{object}{a \code{\link[caret]{train}} object}

\item{validate_for_stacking}{a logical indicating whether to validate the object for stacked predictions}
}
\value{
a character string
}
\description{
Extract the model type from a \code{\link[caret]{train}} object.
For classification, validates that the model can predict probabilities, and,
 if stacked predictions are requested, that classProbs = TRUE.
}
\keyword{internal}
