% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractModelTypes}
\alias{extractModelTypes}
\title{Extracts the model types from a list of train model}
\usage{
extractModelTypes(list_of_models)
}
\arguments{
\item{list_of_models}{an object of class caretList}
}
\description{
Extracts the model types from a list of train model
}

