% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractModFrame}
\alias{extractModFrame}
\title{Extract a dataframe of all predictors used in a caretEnsemble object.}
\usage{
extractModFrame(model)
}
\arguments{
\item{model}{a \code{caretEnsemble} to extract predictors from}
}
\value{
A data.frame combining all of the variables used across all models.
}
\description{
This function constructs a dataframe consisting of the outcome
and all of the predictors used in any of the models ensembled in a \code{caretEnsemble}
object.
}

