% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretEnsemble.R
\name{extractModRes}
\alias{extractModRes}
\title{Extract the model accuracy metrics of the individual models in an ensemble object.}
\usage{
extractModRes(ensemble)
}
\arguments{
\item{ensemble}{a caretEnsemble to make predictions from.}
}
\description{
Extract the model accuracy metrics of the individual models in an ensemble object.
}

