% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extractObsLevels}
\alias{extractObsLevels}
\title{Extract the observed levels from a list of models}
\usage{
extractObsLevels(list_of_models)
}
\arguments{
\item{list_of_models}{an object of class caretList}
}
\description{
Extract the observed levels from a list of models
}
