% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretEnsemble.R
\name{extractPredObsResid}
\alias{extractPredObsResid}
\title{Extract the best predictions and observations from a train object}
\usage{
extractPredObsResid(object, show_class_id = 2L)
}
\arguments{
\item{object}{a \code{train} object}

\item{show_class_id}{For classification only: which class level to use for residuals}
}
\value{
a data.table with predictions, observeds, and residuals
}
\description{
This function extracts the best predictions and observations from a train object
and then calculates residuals.  It only uses one class for classification models, by default class 2.
}
