% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extractTrainingData}
\alias{extractTrainingData}
\title{Extract the training data from a caretList}
\usage{
extractTrainingData(x)
}
\arguments{
\item{x}{a caretList object}
}
\description{
Extract the training data from a caretList
}
