% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.caretEnsemble}
\alias{fortify.caretEnsemble}
\title{Supplement the data fitted to a caret ensemble model with model fit statistics}
\usage{
\method{fortify}{caretEnsemble}(model, data = NULL, ...)
}
\arguments{
\item{data}{a data set, defaults to the data used to fit the model}

\item{...}{additional arguments to pass to fortify}

\item{mode}{a \code{caretEnsemble} to extract predictors from}
}
\value{
The original data with extra columns for fitted values and residuals
}
\description{
This function constructs a dataframe consisting of the outcome,
all of the predictors used in any of the models ensembled in a \code{caretEnsemble}
object, and some model fit statistics.
}

