% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getBinaryTargetLevel}
\alias{getBinaryTargetLevel}
\title{Return the configured target binary class level}
\usage{
getBinaryTargetLevel()
}
\value{
Currently configured binary target level (as integer equal to 1 or 2)
}
\description{
For binary classification problems, ensemble
stacks and certain performance measures require an awareness
of which class in a two-factor outcome is the "target" class.
By default, this class will be assumed to be the first level in
an outcome factor but that setting can be overridden using
\code{setBinaryTargetLevel(2L)}.
}
\seealso{
setBinaryTargetLevel
}
