% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretEnsemble.R
\name{getMetric}
\alias{getMetric}
\title{Extract accuracy metrics from a \code{\link[caret]{train}} model}
\usage{
getMetric(x, metric = NULL, return_sd = FALSE)
}
\arguments{
\item{x}{a train object}

\item{metric}{a character string representing the metric to extract}

\item{return_sd}{a boolean indicating whether to return the SD of the metric instead of the average.}
}
\value{
A numeric representing the metric desired metric.
}
\description{
Extract the cross-validated accuracy metrics or their SDs from caret.
}
