% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getMulticlassExcludedLevel}
\alias{getMulticlassExcludedLevel}
\title{Return the configured multiclass excluded level}
\usage{
getMulticlassExcludedLevel()
}
\value{
Currently configured multiclass excluded level (as integer)
}
\description{
To train a model using probability outputs
provided by other models in a classification problem, it is
necessary to exclude one of the classes. By default, this class
is assumed to be the first level in an outcome factor,
but this setting can be overridden using
\code{setMulticlassTargetLevel(3L)} if the classification
problem has at least 3 classes.
}
\seealso{
setMulticlassTargetLevel
}
