% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretEnsemble.R
\name{is.caretEnsemble}
\alias{is.caretEnsemble}
\title{Check if an object is a caretEnsemble object}
\usage{
is.caretEnsemble(object)
}
\arguments{
\item{object}{an R object}
}
\description{
Check if an object is a caretEnsemble object
}

