% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{is.caretList}
\alias{is.caretList}
\title{Check if an object is a caretList object}
\usage{
is.caretList(object)
}
\arguments{
\item{object}{an R object}
}
\description{
Check if an object is a caretList object
}

