% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretStack.R
\name{is.caretStack}
\alias{is.caretStack}
\title{Check if an object is a caretStack object}
\usage{
is.caretStack(object)
}
\arguments{
\item{object}{an R object}
}
\description{
Check if an object is a caretStack object
}

