% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationImportance.R
\name{isClassifier}
\alias{isClassifier}
\title{Is Classifier}
\usage{
isClassifier(model)
}
\arguments{
\item{model}{A train object from the caret package.}
}
\value{
A logical indicating whether the model is a classifier.
}
\description{
Check if a model is a classifier.
}
\keyword{internal}
