% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationImportance.R
\name{mae}
\alias{mae}
\title{Compute MAE}
\usage{
mae(a, b)
}
\arguments{
\item{a}{A numeric vector.}

\item{b}{A numeric vector.}
}
\value{
A numeric scalar.
}
\description{
Compute the mean absolute error between two vectors.
}
\keyword{internal}
