% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{makePredObsMatrix}
\alias{makePredObsMatrix}
\title{Make a prediction matrix from a list of models}
\usage{
makePredObsMatrix(list_of_models)
}
\arguments{
\item{list_of_models}{an object of class caretList}
}
\description{
Extract obs from one models, and a matrix of predictions from all other models, a
helper function
}

