% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{models.class}
\alias{models.class}
\title{caretList of classification models}
\description{
Data for the caretEnsemble package
}
\author{
Zachary Deane-Mayer \email{zach.mayer@gmail.com}
}
\keyword{data}
\keyword{internal}
