% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{models.reg}
\alias{models.reg}
\title{caretList of regression models}
\description{
caretList of regression models
}
\author{
Zachary Deane-Mayer \email{zach.mayer@gmail.com}
}
\keyword{data}
\keyword{internal}
