% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multiPredict}
\alias{multiPredict}
\title{Create a matrix of predictions for each of the models in a list}
\usage{
multiPredict(list_of_models, type, ...)
}
\arguments{
\item{list_of_models}{a list of caret models to make predictions for}

\item{type}{Classification or Regression}

\item{...}{additional arguments to pass to predict.train. Pass the \code{newdata}
argument here, DO NOT PASS the "type" argument.  Classification models will
return probabilities if possible, and regression models will return "raw".}
}
\description{
Make a matrix of predictions from a list of caret models
}

