% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caretEnsemble.R
\name{multiResiduals}
\alias{multiResiduals}
\title{Calculate the residuals from all component models of a caretEnsemble.}
\usage{
multiResiduals(object, ...)
}
\arguments{
\item{object}{a \code{caretEnsemble} to make predictions from.}

\item{...}{other arguments to be passed to residuals}
}
\value{
A data.frame in the long format with columns for the model method,
the observation id, yhat for the fitted values, resid for the residuals, and
y for the observed value.
}
\description{
This function calculates raw residuals for both regression and
classification \code{train} objects within a \code{\link{caretEnsemble}}.
}

