% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationImportance.R
\name{normalize_to_one}
\alias{normalize_to_one}
\title{Normalize to One}
\usage{
normalize_to_one(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
Normalize a vector to sum to one.
}
\keyword{internal}
