% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{plot.caretList}
\alias{plot.caretList}
\title{Plot a caretList object}
\usage{
\method{plot}{caretList}(x, metric = NULL, ...)
}
\arguments{
\item{x}{a caretList object}

\item{metric}{which metric to plot}

\item{...}{ignored}
}
\value{
A ggplot2 object
}
\description{
This function plots the performance of each model in a caretList object.
}
