% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretStack.R
\name{plot.caretStack}
\alias{plot.caretStack}
\title{Plot a caretStack object}
\usage{
\method{plot}{caretStack}(x, metric = NULL, ...)
}
\arguments{
\item{x}{a caretStack object}

\item{metric}{which metric to plot.  If NULL, will use the default metric used to train the model.}

\item{...}{ignored}
}
\value{
a ggplot2 object
}
\description{
This function plots the performance of each model in a caretList object.
}
