% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable_importance.R
\name{plot_group}
\alias{plot_group}
\title{Plot a group of variable importances}
\usage{
plot_group(imp_dt, title, fill_colors, y_max)
}
\arguments{
\item{imp_dt}{data.table with columns `method`, `weight`, `is_stat`.}

\item{title}{Title for the plot.}

\item{fill_colors}{Named vector: `FALSE` = normal bar color, `TRUE` = gray for statistic bar.}

\item{y_max}{Numeric maximum for the y-axis.}
}
\value{
ggplot object.
}
\description{
Generates a ggplot bar chart for a given set of variable importances.
}
