% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedyOpt.R
\name{predict.greedyMSE}
\alias{predict.greedyMSE}
\title{Predict method for greedyMSE}
\usage{
\method{predict}{greedyMSE}(object, newdata, ...)
}
\arguments{
\item{object}{A greedyMSE object.}

\item{newdata}{A numeric matrix of new data.}

\item{...}{Additional arguments. Ignored.}
}
\value{
A numeric matrix of predictions.
}
\description{
Predict method for greedyMSE objects.
}
