% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable_importance.R
\name{prepare_importance}
\alias{prepare_importance}
\title{Prepare variable importance data.table from a caretStack}
\usage{
prepare_importance(stack_model, newdata)
}
\arguments{
\item{stack_model}{A trained `caretStack` model.}

\item{newdata}{A data frame used for calculating variable importance.}
}
\value{
`data.table` with columns `method` (variable names) and `weight` (importance).
}
\description{
Extracts variable importance from a `caretStack` model and returns a
`data.table` with columns `method` and `weight` sorted by importance.
}
