% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.caretStack}
\alias{print.caretStack}
\title{Print a caretStack object}
\usage{
\method{print}{caretStack}(x, ...)
}
\arguments{
\item{x}{An object of class caretStack}

\item{...}{ignored}
}
\description{
This is a function to print a caretStack.
}

