% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedyOpt.R
\name{print.greedyMSE}
\alias{print.greedyMSE}
\title{Print method for greedyMSE}
\usage{
\method{print}{greedyMSE}(x, ...)
}
\arguments{
\item{x}{A greedyMSE object.}

\item{...}{Additional arguments. Ignored.}
}
\description{
Print method for greedyMSE objects.
}
