% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{print.summary.caretList}
\alias{print.summary.caretList}
\title{Print a summary.caretList object}
\usage{
\method{print}{summary.caretList}(x, ...)
}
\arguments{
\item{x}{An object of class summary.caretList}

\item{...}{ignored}
}
\description{
This is a function to print a summary.caretList
}
