% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretStack.R
\name{print.summary.caretStack}
\alias{print.summary.caretStack}
\title{Print a summary.caretStack object}
\usage{
\method{print}{summary.caretStack}(x, ...)
}
\arguments{
\item{x}{An object of class summary.caretStack}

\item{...}{ignored}
}
\description{
This is a function to print a summary.caretStack.
}
