% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretStack.R
\name{process_which_var}
\alias{process_which_var}
\title{Process which.var}
\usage{
process_which_var(use_original_features, feature_colnames, which.var)
}
\arguments{
\item{use_original_features}{a logical indicating whether to include the original features in the stack.}

\item{feature_colnames}{a character vector of the names of the original features used to train the stack.}

\item{which.var}{a character vector of the names of the original features to include in the stack or a numeric
vector with the indexes of the original features to include in the stack. If NULL, will include all original
features. If use_original_features is FALSE, this argument is ignored.}
}
\description{
Process the which.var argument for caretStack.
}
\keyword{internal}
