% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qpOptRMSE}
\alias{qpOptRMSE}
\title{Quadratic optimization of the reduced mean square error}
\usage{
qpOptRMSE(x, y, ...)
}
\arguments{
\item{X}{the matrix of predictors}

\item{Y}{the dependent variable}
}
\value{
A numeric of the weights for each model
}
\description{
This algorithm optimizes the RMSE for regression models to avoid convex combinations
}

