% Generated by roxygen2 (4.0.2): do not edit by hand
\name{residuals.caretEnsemble}
\alias{residuals.caretEnsemble}
\title{Calculate the residuals from a caretEnsemble.}
\usage{
\method{residuals}{caretEnsemble}(object, ...)
}
\arguments{
\item{object}{a \code{caretEnsemble} to make predictions from.}

\item{...}{other arguments to be passed to residuals}
}
\value{
A numeric of the residuals.
}
\description{
This function calculates raw residuals for both regression and
classification \code{caretEnsemble} objects.
}

