% Generated by roxygen2 (4.0.2): do not edit by hand
\name{safeOptAUC}
\alias{safeOptAUC}
\title{Safe optimization of the AUC}
\usage{
safeOptAUC(X, Y, iter = 100L)
}
\arguments{
\item{X}{the matrix of predictors}

\item{Y}{the dependent variable}

\item{iter}{an integer for the number of iterations}
}
\value{
A numeric of the weights for each model
}
\description{
This algorithm optimizes the AUC for regression models to avoid ensembling
where the ensembled model fits worse than any component model
}

