% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretStack.R
\name{set_excluded_class_id}
\alias{set_excluded_class_id}
\title{Set excluded class id}
\usage{
set_excluded_class_id(object, model_type)
}
\arguments{
\item{object}{a caretStack object}

\item{model_type}{the model type as a character vector with length 1}
}
\description{
Set the excluded class id for a caretStack object
}
\keyword{internal}
