% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{setBinaryTargetLevel}
\alias{setBinaryTargetLevel}
\title{Set the target binary class level}
\usage{
setBinaryTargetLevel(level)
}
\arguments{
\item{level}{an integer in \{1, 2\} to be used as target outcome level}
}
\description{
For binary classification problems, ensemble
stacks and certain performance measures require an awareness
of which class in a two-factor outcome is the "target" class.
By default, the first level in an outcome factor is used but
this value can be overridden using \code{setBinaryTargetLevel(2L)}
}
\seealso{
getBinaryTargetLevel
}

