% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{setMulticlassExcludedLevel}
\alias{setMulticlassExcludedLevel}
\title{Set the multiclass excluded level}
\usage{
setMulticlassExcludedLevel(level)
}
\arguments{
\item{level}{an integer to be used as excluded}
}
\description{
To train a model using probability outputs
provided by other models in a classification problem, it is
necessary to exclude one of the classes. By default, this class
is assumed to be the first level in an outcome factor,
but this setting can be overridden using
\code{setMulticlassTargetLevel(3L)} if the classification
problem has at least 3 classes.
}
\note{
Setting this value outside the range between 1 and
the number of classes will cause caretStack to train the model
with the probabilities associated with ALL classes, leading to
potential collinearity issues.
}
\seealso{
getMulticlassExcludedLevel
}
