% Generated by roxygen2 (4.0.0): do not edit by hand
\name{setSeeds}
\alias{setSeeds}
\title{Create uniform seeds across model fits to allow for ensembling}
\usage{
setSeeds(ctrl, M)
}
\arguments{
\item{ctrl}{a \code{\link{trainControl}} object passed by the user}

\item{M}{the maximum number of resamples necessary}
}
\value{
A \code{\link{trainControl}} object with a new slot, seeds
}
\description{
Create uniform seeds across model fits to allow for ensembling
}

