% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationImportance.R
\name{shuffled_mae}
\alias{shuffled_mae}
\title{Shuffled MAE}
\usage{
shuffled_mae(model, original_data, target, pred_type, shuffle_idx)
}
\arguments{
\item{original_data}{A data.table of the original data.}

\item{target}{A matrix of target values.}

\item{shuffle_idx}{A vector of shuffled indices.}
}
\value{
A numeric vector of mean absolute errors.
}
\description{
Compute the mean absolute error of a model's predictions when a variable is shuffled.
}
\keyword{internal}
