% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretStack.R
\name{stackedTrainResiduals}
\alias{stackedTrainResiduals}
\title{Extracted stacked residuals for the autoplot}
\usage{
stackedTrainResiduals(object, show_class_id = 2L)
}
\arguments{
\item{object}{a \code{train} object}

\item{show_class_id}{For classification only: which class level to use for residuals}
}
\value{
a data.table::data.table with predictions, observeds, and residuals
}
\description{
This function extracts the predictions, observeds, and residuals from a \code{train} object.
It uses the object's stacked predictions from cross-validation.
}
\keyword{internal}
