% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{summary,caretEnsemble-method}
\alias{summary,caretEnsemble-method}
\title{Summarize the results of caretEnsemble for the user.}
\usage{
\S4method{summary}{caretEnsemble}(object, ...)
}
\arguments{
\item{object}{a \code{\link{caretEnsemble}} to make predictions from.}

\item{...}{optional additional parameters.}
}
\description{
Summarize the results of caretEnsemble for the user.
}

