\name{summary.caretEnsemble}
\alias{summary.caretEnsemble}
\title{Summarize the results of caretEnsemble for the user.}
\usage{
\method{summary}{caretEnsemble}(ensemble)
}
\arguments{
  \item{ensemble}{a \code{\linkS4class{caretEnsemble}} to
  make predictions from.}
}
\description{
Summarize the results of caretEnsemble for the user.
}

