% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{summary.caretList}
\alias{summary.caretList}
\title{Summarize a caretList}
\usage{
\method{summary}{caretList}(object, metric = NULL, ...)
}
\arguments{
\item{object}{a caretList object}

\item{metric}{The metric to show.  If NULL will use the metric used to train each model}

\item{...}{passed to extractMetric}
}
\value{
A data.table with metrics from each model.
}
\description{
This function summarizes the performance of each model in a caretList object.
}
