% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.caretStack}
\alias{summary.caretStack}
\title{Summarize a caretStack object}
\usage{
\method{summary}{caretStack}(object, ...)
}
\arguments{
\item{object}{An object of class caretStack}

\item{...}{ignored}
}
\description{
This is a function to summarize a caretStack.
}

