% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tmpExtract}
\alias{tmpExtract}
\title{Helper function to extract tune grids from arguments passed to tuneList}
\usage{
tmpExtract(x)
}
\arguments{
\item{x}{list of tuning parameters for train objects}
}
\value{
The maximum length of any tuneGrid in the list
}
\description{
Returns the tuneGrids of the tuneList
}

