% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{trControlCheck}
\alias{trControlCheck}
\title{Check that the trainControl object supplied by the user is valid and has defined re-sampling indexes.}
\usage{
trControlCheck(x, y)
}
\arguments{
\item{x}{a trainControl object.}

\item{y}{the target for the model.  Used to determine resampling indexes.}
}
\description{
This function checks the user-supplied trainControl object and makes sure it has all the required fields.  If the resampling indexes are missing, it adds them to the model.  If savePredictions=FALSE or "none", this function sets it to "final".
}
