% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tuneExtract}
\alias{tuneExtract}
\title{Extract the tuning parameters from a list provided by the user}
\usage{
tuneExtract(x)
}
\arguments{
\item{x}{a list of tuning parameters and methods passed by the user}
}
\value{
A simple list of the tuning parameters
}
\description{
Extract the tuning parameters from a list provided by the user
}

