% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{validateBinaryTargetLevel}
\alias{validateBinaryTargetLevel}
\title{Validate arguments given as binary target level}
\usage{
validateBinaryTargetLevel(arg)
}
\arguments{
\item{arg}{argument to potentially be used as new target level}
}
\value{
Binary target level (as integer equal to 1 or 2)
}
\description{
Helper function used to ensure that target
binary class levels given by clients can be coerced to an integer
and that the resulting integer is in \{1, 2\}.
}
