% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{validateExcludedClass}
\alias{validateExcludedClass}
\title{Validate the excluded class}
\usage{
validateExcludedClass(arg)
}
\arguments{
\item{arg}{The value to check}
}
\value{
integer
}
\description{
Helper function to ensure that the excluded level for classification is an integer.
Set to 0L to exclude no class.
}
