% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{validateMethodList}
\alias{validateMethodList}
\title{Check that the tuning parameters list supplied by the user is valid}
\usage{
validateMethodList(methodList)
}
\arguments{
\item{x}{a list of user-supplied tuning parameters and methods}
}
\value{
a named list of caretModelSpec objects
}
\description{
This function makes sure the tuning parameters passed by the user
are valid and have the proper naming, etc.
}
\keyword{internal}
