% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{validateMulticlassExcludedLevel}
\alias{validateMulticlassExcludedLevel}
\title{Validate arguments given as multiclass excluded level}
\usage{
validateMulticlassExcludedLevel(arg)
}
\arguments{
\item{arg}{argument to potentially be used as new excluded level}
}
\value{
Multiclass excluded level (as integer)
}
\description{
Helper function used to ensure that excluded
multiclass levels given by clients can be coerced to an integer.
}
