% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretStack.R
\name{varImp.caretStack}
\alias{varImp.caretStack}
\title{Variable importance for caretStack}
\usage{
\method{varImp}{caretStack}(object, newdata = NULL, normalize = TRUE, ...)
}
\arguments{
\item{object}{An object of class caretStack}

\item{newdata}{the data to use for computing importance. If NULL, will use the stacked predictions from the models}

\item{normalize}{a logical indicating whether to normalize the importances to sum to one.}

\item{...}{passed to predict.caretList}
}
\description{
This is a function to extract variable importance from a caretStack.
}
