% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedyOpt.R
\name{varImp.greedyMSE}
\alias{varImp.greedyMSE}
\title{variable importance for a greedyMSE model}
\usage{
\method{varImp}{greedyMSE}(object, ...)
}
\arguments{
\item{object}{A greedyMSE object.}

\item{...}{Additional arguments. Ignored.}
}
\description{
Variable importance for a greedyMSE model.
}
