% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wtd.sd}
\alias{wtd.sd}
\title{Calculate a weighted standard deviation}
\usage{
wtd.sd(x, weights = NULL, normwt = FALSE, na.rm = TRUE)
}
\arguments{
\item{x}{a vector of numerics}

\item{weights}{a vector of weights equal to length of x}

\item{normwt}{a logical indicating whether the weights should be normalized to 1}

\item{na.rm}{a logical indicating how to handle missing values}
}
\description{
Used to weight deviations among ensembled model preditions
}

