% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{X.class}
\alias{X.class}
\title{data for classification}
\description{
data for classification
}
\author{
Zachary Deane-Mayer \email{zach.mayer@gmail.com}
}
\keyword{data}
\keyword{internal}
