% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{as.caretList}
\alias{as.caretList}
\title{Convert object to caretList object}
\usage{
as.caretList(object)
}
\arguments{
\item{object}{R Object}
}
\value{
a \code{\link{caretList}} object
}
\description{
Converts object into a caretList
}
