% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.caretEnsemble}
\alias{autoplot.caretEnsemble}
\title{Convenience function for more in-depth diagnostic plots of caretEnsemble objects}
\usage{
\method{autoplot}{caretEnsemble}(object, which = c(1:6), mfrow = c(2, 3),
  xvars = NULL, ...)
}
\arguments{
\item{object}{a \code{caretEnsemble} object}

\item{which}{an integer index for which of the plots to print}

\item{mfrow}{an integer vector of length 2 specifying the number of rows and columns for plots}

\item{xvars}{a vector of the names of x variables to plot against residuals}

\item{...}{additional arguments to pass to autoplot}
}
\value{
A grid of diagnostic plots
}
\description{
This function provides a more robust series of diagnostic plots
for a caretEnsemble object.
}

