% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S3GenericExtenstions.R
\name{c.caretList}
\alias{c.caretList}
\title{S3 definition for concatenating models}
\usage{
\method{c}{caretList}(...)
}
\arguments{
\item{modelList1}{an object of class caretList or train}

\item{additionalModels}{an object of class caretList or train}
}
\value{
a \code{\link{caretList}} object
}
\description{
take two lists of models and concatenat them for future Ensamble'ing
}
\examples{
\dontrun{
 model_list1 <- caretList(Class ~ .,
  data=Sonar, trControl = ctrl1,
  tuneList = list(
                 glm=caretModelSpec(method='glm', family='binomial'),
                 rpart=caretModelSpec(method='rpart')
                 ),
   metric='ROC')

model_list2 <- caretList(Class ~ .,
                         data=Sonar,
                         trControl = ctrl1,
                         tuneList = list(
                           glm=caretModelSpec(method='rpart'),
                           rpart=caretModelSpec(method='rf')
                         ),
                         metric='ROC')

 bigList <- c(model_list1, model_list2)
}
}

