% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{caretEnsemble}
\alias{caretEnsemble}
\title{Class "caretEnsemble" of ensembled train objects from the caret package}
\usage{
caretEnsemble(all.models, optFUN = NULL, ...)

caretEnsemble(all.models, optFUN = NULL, ...)
}
\arguments{
\item{all.models}{a list of caret models to ensemble.}

\item{optFUN}{the optimization function to use}

\item{...}{additional arguments to pass to the optimization function}
}
\value{
a \code{\link{caretEnsemble}} object
}
\description{
Ensembled model from input train objects.

Find a good linear combination of several classification or regression models,
using either linear regression, elastic net regression, or greedy optimization.
}
\details{
The object has the following items
\itemize{
\item{models - a list of the original models to be ensembled}
\item{weights - a matrix with the weights for each model}
\item{error - the final accuracy metric of the ensembled models}
}

Every model in the "library" must be a separate \code{train} object.  For
example, if you wish to combine a random forests with several different
values of mtry, you must build a model for each value of mtry.  If you
use several values of mtry in one train model, (e.g. tuneGrid =
expand.grid(.mtry=2:5)), caret will select the best value of mtry
before we get a chance to include it in the ensemble.  By default,
RMSE is used to ensemble regression models, and AUC is used to ensemble
Classification models.  This function does not currently support multi-class
problems
}
\section{Objects from the Class}{
 Objects are created by calls to
\code{\link{caretEnsemble}}.
}
\references{
\url{http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.60.2859&rep=rep1&type=pdf}
}
\seealso{
\code{\link{caretEnsemble}}
}
\keyword{classes}

