% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{caretList}
\alias{caretList}
\title{Create a list of several train models from the caret package}
\usage{
caretList(
  ...,
  metric = NULL,
  trControl = NULL,
  methodList = NULL,
  tuneList = NULL,
  continue_on_fail = FALSE
)
}
\arguments{
\item{...}{arguments to pass to \code{\link[caret]{train}}.
These arguments will determine which train method gets dispatched.}

\item{metric}{a string, the metric to optimize for.  If NULL, we will choose a good one.}

\item{trControl}{a \code{\link[caret]{trainControl}} object.  If null, we will construct a good one.}

\item{methodList}{optional, a character vector of caret models to ensemble.
One of methodList or tuneList must be specified.}

\item{tuneList}{optional, a NAMED list of caretModelSpec objects.
This much more flexible than methodList and allows the
specification of model-specific parameters (e.g. passing trace=FALSE to nnet)}

\item{continue_on_fail, }{logical, should a valid caretList be returned that
excludes models that fail, default is FALSE}
}
\value{
A list of \code{\link[caret]{train}} objects. If the model fails to build,
it is dropped from the list.
}
\description{
Build a list of train objects suitable for ensembling using the \code{\link{caretEnsemble}}
function.
}
\examples{
\dontrun{
myControl <- trainControl(method = "cv", number = 5)
caretList(
  Sepal.Length ~ Sepal.Width,
  head(iris, 50),
  methodList = c("glm", "lm"),
  trControl = myControl
)
caretList(
  Sepal.Length ~ Sepal.Width,
  head(iris, 50),
  methodList = c("lm"),
  tuneList = list(
    nnet = caretModelSpec(method = "nnet", trace = FALSE, tuneLength = 1)
  ),
  trControl = myControl
)
}
}
