% Generated by roxygen2 (4.0.2): do not edit by hand
\name{caretModelSpec}
\alias{caretModelSpec}
\title{Generate a specification for fitting a caret model}
\usage{
caretModelSpec(method = "rf", ...)
}
\arguments{
\item{method}{the modeling method to pass to caret::train}

\item{args}{the arguments to pass to the model}
}
\value{
a list of lists
}
\description{
A caret model specificaiton consists of 2 parts: a model (as a string) and the argments to the train call for fitting that model
}

