% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_bestpreds_indexes}
\alias{check_bestpreds_indexes}
\title{Check row indexes}
\usage{
check_bestpreds_indexes(modelLibrary)
}
\arguments{
\item{modelLibrary}{a list of predictions from caret models}
}
\description{
Check that the row indexes from a caretList are valid
}
