% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_bestpreds_obs}
\alias{check_bestpreds_obs}
\title{Check observeds}
\usage{
check_bestpreds_obs(modelLibrary)
}
\arguments{
\item{modelLibrary}{a list of predictins from caret models}
}
\description{
Check that a list of observed values from a caretList are valid
}

