% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_bestpreds_preds}
\alias{check_bestpreds_preds}
\title{Check predictions}
\usage{
check_bestpreds_preds(modelLibrary)
}
\arguments{
\item{modelLibrary}{a list of predictins from caret models}
}
\description{
Check that a list of predictions from a caretList are valid
}

