% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_caretList_model_types}
\alias{check_caretList_model_types}
\title{Checks that caretList models are all of the same type.}
\usage{
check_caretList_model_types(list_of_models)
}
\arguments{
\item{list_of_models}{a list of caret models to check}
}
\description{
Checks that caretList models are all of the same type.
}

