% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretEnsemble-package.R
\docType{data}
\name{models.class}
\alias{models.class}
\alias{models.reg}
\alias{X.class}
\alias{Y.class}
\alias{X.reg}
\alias{Y.reg}
\title{caretList of classification models}
\description{
Data for the caretEnsemble package
}
\author{
Zachary Deane-Mayer \email{zach.mayer@gmail.com}
}
\keyword{data}
