% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretPredict.R
\name{extractBestPreds}
\alias{extractBestPreds}
\title{Extract the best predictions from a train object}
\usage{
extractBestPreds(x, aggregate_resamples = TRUE)
}
\arguments{
\item{x}{a train object}

\item{aggregate_resamples}{logical, whether to aggregate resamples by keys. Default is TRUE.}
}
\value{
a data.table::data.table with predictions
}
\description{
Extract the best predictions from a train object.
}
\keyword{internal}
