\name{extractBestPreds}
\alias{extractBestPreds}
\title{Extract predictions for the best tune from a list of caret models}
\usage{
  extractBestPreds(list_of_models)
}
\arguments{
  \item{list_of_models}{a list of caret models to extract
  predictions from}
}
\description{
  Extract predictions for the best tune from a list of
  caret models
}

