% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretPredict.R
\name{extractBestPreds}
\alias{extractBestPreds}
\title{Extract the best predictions from a train object}
\usage{
extractBestPreds(x)
}
\arguments{
\item{x}{a train object}
}
\value{
a data.table::data.table with predictions
}
\description{
Extract the best predictions from a train object.
}
\keyword{internal}
