% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extractBestPreds}
\alias{extractBestPreds}
\title{Extract the best predictions from a list of train objects}
\usage{
extractBestPreds(list_of_models)
}
\arguments{
\item{list_of_models}{an object of class caretList or a list of caret models}
}
\description{
Extract predictions for the best tune from a list of caret models
}
