% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extractModelType}
\alias{extractModelType}
\title{Extracts the model type from a list of train models}
\usage{
extractModelType(list_of_models)
}
\arguments{
\item{list_of_models}{an object of class caretList}
}
\description{
Extracts the model types from a list of train models
}
