% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extractModelTypes}
\alias{extractModelTypes}
\title{Extracts the model types from a list of train model}
\usage{
extractModelTypes(list_of_models)
}
\arguments{
\item{list_of_models}{an object of class caretList}
}
\description{
Extracts the model types from a list of train model
}
