% Generated by roxygen2 (4.0.0): do not edit by hand
\name{extractModRes}
\alias{extractModRes}
\title{Extract the model accuracy metrics of the individual models in an ensemble object.}
\usage{
extractModRes(ensemble)
}
\arguments{
\item{ensemble}{a caretEnsemble to make predictions from.}
}
\description{
Extract the model accuracy metrics of the individual models in an ensemble object.
}

