\name{makePredObsMatrix}
\alias{makePredObsMatrix}
\title{Extract obs from one models, and a matrix of predictions from all other models}
\usage{
  makePredObsMatrix(list_of_models)
}
\arguments{
  \item{list_of_models}{a list of caret models to extract
  predictions from}
}
\description{
  Extract obs from one models, and a matrix of predictions
  from all other models
}

