% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/caretList.R
\name{methodCheck}
\alias{methodCheck}
\title{Check that the methods supplied by the user are valid caret methods}
\usage{
methodCheck(x)
}
\arguments{
\item{x}{a list of user-supplied tuning parameters and methods}
}
\description{
This function uses modelLookup from caret to ensure the list of methods supplied by the user are all models caret can fit.
}

