% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretEnsemble.R
\name{getMetric}
\alias{getMetric}
\alias{getMetricSD}
\alias{getMetric.train}
\alias{getMetricSD.train}
\title{Extract accuracy metrics from a model}
\usage{
getMetric(x, metric, ...)

getMetricSD(x, metric, ...)

\method{getMetric}{train}(x, metric = NULL, ...)

\method{getMetricSD}{train}(x, metric, ...)
}
\arguments{
\item{x}{a train object}

\item{metric}{which metric to get}

\item{...}{passed through}
}
\value{
A numeric representing the metric desired metric.
}
\description{
Extract accuracy metrics from a model

Extract accuracy metrics SDs from a model

Extract a model accuracy metric from a \code{\link{train}} object.

Extract the standard deviation from resamples for an accuracy metric from a model object.
}
