% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMetricSD}
\alias{getMetricSD}
\title{Extract the standard deviation from resamples for an accuracy metric from
a model object.}
\usage{
getMetricSD(x, metric)
}
\arguments{
\item{x}{an object with model performanc metrics}

\item{metric}{a character, either "RMSE" or "AUC" indicating which metric to extract}
}
\value{
A numeric for the standard deviation of the selected metric across
resamples for the best model
}
\description{
Extract the standard deviation from resamples for an accuracy metric from
a model object.
}

