% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.caretStack}
\alias{plot.caretStack}
\title{Plot a caretStack object}
\usage{
\method{plot}{caretStack}(x, ...)
}
\arguments{
\item{x}{An object of class caretStack}

\item{...}{passed to plot}
}
\description{
This is a function to plot a caretStack.
}

