% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/caretStack.R
\name{plot.caretStack}
\alias{plot.caretStack}
\title{Plot a caretStack object}
\usage{
\method{plot}{caretStack}(x, ...)
}
\arguments{
\item{x}{An object of class caretStack}

\item{...}{passed to plot}
}
\description{
This is a function to plot a caretStack.
}
\examples{
library('rpart')
models <- caretList(
  x=iris[1:100,1:2],
  y=iris[1:100,3],
  trControl=trainControl(method='cv'),
  methodList=c('rpart', 'glm')
)
meta_model <- caretStack(models, method='rpart', tuneLength=2)
plot(meta_model)
}

