% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.caretEnsemble}
\alias{predict.caretEnsemble}
\title{Make predictions from a caretEnsemble. This function passes the data to each function in
turn to make a matrix of predictions, and then multiplies that matrix by the vector of
weights to get a single, combined vector of predictions.}
\usage{
\method{predict}{caretEnsemble}(object, keepNA = TRUE, se = NULL, ...)
}
\arguments{
\item{object}{a \code{\link{caretEnsemble}} to make predictions from.}

\item{keepNA}{a logical indicating whether predictions should be made for all
cases where sufficient data exists or only for complete cases across all models. When
TRUE this does not predict for missing values. When FALSE, missing values are overwritten
with predictions where possible.}

\item{se}{logical, should prediction errors be produced? Default is false.}

\item{...}{arguments (including newdata) to pass to predict.train. These arguments
must be named}
}
\value{
If keepNA is set to true this will return either a vector of numeric predictions
or a data.frame of numeric predictions in the first column with model disagreement in
the second column if \code{se = TRUE}. If keepNA is set to FALSE and missing values
are present in the new data, then a \code{list} is produced. The list contains
the data.frame of predictions in the first slot and a matrix of model weights
for each observation in the second slot.
}
\description{
Make predictions from a caretEnsemble. This function passes the data to each function in
turn to make a matrix of predictions, and then multiplies that matrix by the vector of
weights to get a single, combined vector of predictions.
}

