% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{predict.caretList}
\alias{predict.caretList}
\title{Create a matrix of predictions for each of the models in a caretList}
\usage{
\method{predict}{caretList}(object, newdata = NULL, verbose = FALSE, excluded_class_id = 0L)
}
\arguments{
\item{object}{an object of class caretList}

\item{newdata}{New data for predictions.  It can be NULL, but this is ill-advised.}

\item{verbose}{Logical. If FALSE no progress bar is printed if TRUE a progress
bar is shown. Default FALSE.}

\item{excluded_class_id}{Integer.  The class id to drop when predicting for multiclass}
}
\description{
Make a matrix of predictions from a list of caret models
}
