% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/caretStack.R
\name{print.caretStack}
\alias{print.caretStack}
\title{Print a caretStack object}
\usage{
\method{print}{caretStack}(x, ...)
}
\arguments{
\item{x}{An object of class caretStack}

\item{...}{ignored}
}
\description{
This is a function to print a caretStack.
}
\examples{
library('rpart')
models <- caretList(
  x=iris[1:100,1:2],
  y=iris[1:100,3],
  trControl=trainControl(method='cv'),
  methodList=c('rpart', 'glm')
)
meta_model <- caretStack(models, method='lm')
print(meta_model)
}

