% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qpOptAUC}
\alias{qpOptAUC}
\title{Quadratic optimization of the AUC}
\usage{
qpOptAUC(x, y, offset = NULL, ...)
}
\arguments{
\item{X}{the matrix of predictors}

\item{Y}{the dependent variable}
}
\value{
A numeric of the weights for each model
}
\description{
This algorithm optimizes the AUC for regression models to avoid convex combinations
}

