% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretList.R
\name{[.caretList}
\alias{[.caretList}
\title{Index a caretList}
\usage{
\method{[}{caretList}(object, index)
}
\arguments{
\item{object}{an object of class caretList}

\item{index}{selected index}
}
\description{
Index a caret list to extract caret models into a new caretList object
}
