% Generated by roxygen2 (4.0.0): do not edit by hand
\name{summary.caretEnsemble}
\alias{summary.caretEnsemble}
\title{Summarize the results of caretEnsemble for the user.}
\usage{
\method{summary}{caretEnsemble}(object, ...)
}
\arguments{
\item{object}{a \code{\link{caretEnsemble}} to make predictions from.}

\item{...}{optional additional parameters.}
}
\description{
Summarize the results of caretEnsemble for the user.
}

