\name{summary.caretEnsemble}
\alias{summary.caretEnsemble}
\title{Summarize the results of caretEnsemble for the user.}
\usage{
\method{summary}{caretEnsemble}(object, ...)
}
\arguments{
  \item{object}{a \code{\linkS4class{caretEnsemble}} to
  make predictions from.}

  \item{...}{optional additional parameters.}
}
\description{
Summarize the results of caretEnsemble for the user.
}

