% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretEnsemble.R
\name{summary.caretEnsemble}
\alias{summary.caretEnsemble}
\title{Summarize the results of caretEnsemble for the user.}
\usage{
\method{summary}{caretEnsemble}(object, ...)
}
\arguments{
\item{object}{a \code{\link{caretEnsemble}} to make predictions from.}

\item{...}{optional additional parameters.}
}
\description{
Summarize a caretEnsemble
}
\examples{
\dontrun{
set.seed(42)
models <- caretList(iris[1:50,1:2], iris[1:50,3], methodList=c("glm", "lm"))
ens <- caretEnsemble(models)
summary(ens)
}
}

