% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.caretStack}
\alias{summary.caretStack}
\title{Summarize a caretStack object}
\usage{
\method{summary}{caretStack}(object, ...)
}
\arguments{
\item{object}{An object of class caretStack}

\item{...}{ignored}
}
\description{
This is a function to summarize a caretStack.
}
\examples{
library('rpart')
models <- caretList(
  x=iris[1:100,1:2],
  y=iris[1:100,3],
  trControl=trainControl(method='cv'),
  methodList=c('rpart', 'glm')
)
meta_model <- caretStack(models, method='lm')
summary(meta_model)
}

