% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/caretList.R
\name{tuneCheck}
\alias{tuneCheck}
\title{Check that the tuning parameters list supplied by the user is valid}
\usage{
tuneCheck(x)
}
\arguments{
\item{x}{a list of user-supplied tuning parameters and methods}
}
\description{
This function makes sure the tuning parameters passed by the user are valid and have the proper naming, etc.
}

