% Generated by roxygen2 (4.0.0): do not edit by hand
\name{varImp.caretEnsemble}
\alias{varImp.caretEnsemble}
\title{Calculate the variable importance of variables in a caretEnsemble.}
\usage{
\method{varImp}{caretEnsemble}(x)
}
\arguments{
\item{x}{a \code{caretEnsemble} to make predictions from.}
}
\description{
Calculate the variable importance of variables in a caretEnsemble.
}

