% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{wtd.sd}
\alias{wtd.sd}
\title{Calculate a weighted standard deviation}
\usage{
wtd.sd(x, w = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of numerics}

\item{w}{a vector of weights equal to length of x}

\item{na.rm}{a logical indicating how to handle missing values, default = FALSE}
}
\description{
Used to weight deviations among ensembled model preditions
}
